requirements.txt
Django~=5.0
python-dateutil~=2.9.0.post0